Migrate to Nord Stage 3 song presets
Worth noting is that each "song preset" really consists of 5 program change slots, meaning that first song preset i PC 0, second preset is PC 5, and so on. A complete program change for the Nord Stage 3 should consist of a Bank Select MSB and LSB, and finally the Program Change message.

https://www.norduserforum.com/nord-stage-forum-f3/changing-ns3-programs-via-midi-using-imidipatchbay-t14611.html